feat: global auth setted.

+++ backend/src/main/java/com/gamux/website_api/services/ImageService.java
@@ -62,6 +62,7 @@ public class ImageService {
 
     public void deleteImage(String imgName) throws Exception {
         try {
+            if (imgName == null) return;
             Path filepath = storagePath.resolve(imgName).normalize();
             Files.deleteIfExists(filepath);
         } catch (IOException e) {
